pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "retico-conversational-agent"
authors = [{ name = "Marius Le Chapelier", email = "mariuslechapelier@gmail.com" }]
description = ""
readme = "README.md"
license = { file = "LICENSE" }
version = "0.0.1"
requires-python = ">=3.11.7"
dependencies = [
    "transitions",
    "pydub",
    "webrtcvad",
    "faster_whisper",
    "llama-cpp-python",
    "retico_core @ git+https://github.com/articulab/retico-core.git",
    "retico-wozmic @ git+https://github.com/articulab/retico-wozmic.git",
    "retico-amq @ git+https://github.com/articulab/retico-amq.git",
    # "TTS[py3117] @ git+https://github.com/articulab/CoquiTTS.git@dev",
    "TTS[versionless] @ git+https://github.com/articulab/CoquiTTS",
]
